✒️Feat: UI 분리
